+++ packages/stream_video/lib/src/call/call.dart
@@ -812,6 +812,7 @@ class Call {
       }
     }
 
+    _session?.trace('call.accept', null);
     final result = await _coordinatorClient.acceptCall(cid: state.callCid);
     if (result is Success<None>) {
       _stateManager.lifecycleCallAccepted();
@@ -825,6 +826,7 @@ class Call {
     final state = this.state.value;
     _logger.i(() => '[reject] reason: $reason');
 
+    _session?.trace('call.reject', reason?.value);
     final result = await _coordinatorClient.rejectCall(
       cid: state.callCid,
       reason: reason?.value,