+++ packages/stream_video/lib/src/call/session/call_session.dart
@@ -512,6 +512,7 @@ class CallSession extends Disposable {
   void leave({String? reason}) {
     _logger.d(() => '[leave] reason: $reason');
     _isLeavingOrClosed = true;
+    _tracer.trace('call.leave', reason);
     sfuWS.leave(sessionId: sessionId, reason: reason);
   }